Lab 5.1/pyproject.toml
[project]
name = "lab-5-1"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "pycryptodome>=3.23.0",
]
